Decouple input registration logic from the rest
